+++ build.gradle
@@ -22,8 +22,10 @@ plugins {
 
 sourceCompatibility = 11
 
-group 'org.masjidku'
-version '1.0'
+allprojects {
+    group 'org.masjidku'
+    version '1.0'
+}
 
 repositories {
     mavenCentral()
@@ -45,8 +47,6 @@ dependencies {
 
     // Unit Test
     testCompile group: 'junit', name: 'junit', version: '4.12'
-    testCompile group: 'org.hamcrest', name: 'hamcrest', version: '2.1'
-    testCompile group: 'org.assertj', name: 'assertj-core', version: '3.13.2'
 
     // Hash Security /and other tools
     implementation group: 'com.google.guava', name: 'guava', version: '30.1.1-jre'
