--- src/test/java/org/masjidku/model/UserTest.java
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2021. Creative Commons Legal Code
- *
- *                            CC0 1.0 Universal
- *
- *                                CREATIVE COMMONS CORPORATION IS NOT A LAW FIRM AND DOES NOT PROVIDE
- *                                LEGAL SERVICES. DISTRIBUTION OF THIS DOCUMENT DOES NOT CREATE AN
- *                                ATTORNEY-CLIENT RELATIONSHIP. CREATIVE COMMONS PROVIDES THIS
- *                                INFORMATION ON AN "AS-IS" BASIS. CREATIVE COMMONS MAKES NO WARRANTIES
- *                                REGARDING THE USE OF THIS DOCUMENT OR THE INFORMATION OR WORKS
- *                                PROVIDED HEREUNDER, AND DISCLAIMS LIABILITY FOR DAMAGES RESULTING FROM
- *                                THE USE OF THIS DOCUMENT OR THE INFORMATION OR WORKS PROVIDED
- *                                HEREUNDER.
- */
-
-package org.masjidku.model;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.masjidku.model.user.User;
-
-import static org.junit.Assert.assertEquals;
-
-public class UserTest {
-    private User user;
-
-    @Before
-    public void setUp() {
-        user = new User("ucok",
-                "Ucok",
-                "Ketua",
-                "Aktif",
-                "22 Januari 2020",
-                "21 Januari 2020"
-        );
-    }
-
-    @Test
-    public void getJabatan() {
-        assertEquals("ketua", user.getJabatan().toString);
-    }
-}
