--- src/test/java/org/masjidku/util/DatabaseConnectionTest.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2021. Creative Commons Legal Code
- *
- *                            CC0 1.0 Universal
- *
- *                                CREATIVE COMMONS CORPORATION IS NOT A LAW FIRM AND DOES NOT PROVIDE
- *                                LEGAL SERVICES. DISTRIBUTION OF THIS DOCUMENT DOES NOT CREATE AN
- *                                ATTORNEY-CLIENT RELATIONSHIP. CREATIVE COMMONS PROVIDES THIS
- *                                INFORMATION ON AN "AS-IS" BASIS. CREATIVE COMMONS MAKES NO WARRANTIES
- *                                REGARDING THE USE OF THIS DOCUMENT OR THE INFORMATION OR WORKS
- *                                PROVIDED HEREUNDER, AND DISCLAIMS LIABILITY FOR DAMAGES RESULTING FROM
- *                                THE USE OF THIS DOCUMENT OR THE INFORMATION OR WORKS PROVIDED
- *                                HEREUNDER.
- */
-
-package org.masjidku.util;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-public class DatabaseConnectionTest {
-    DatabaseConnection connection = new DatabaseConnection();
-
-    @Test
-    public void getConnection() {
-        Assert.assertNotNull("Periksa koneksi database!", "");
-    }
-}
